Moved analytical solver function from utils to the test-file
